updated profile page to use react router
