+++ apps/frontend/src/app/Profile/Profile.module.scss
@@ -10,7 +10,7 @@
 
   .sidebar {
     position: absolute;
-    top: 0;
+    top: 80px;
     left: -182px;
     width: 150px;
     color: var(--label-color);
@@ -53,56 +53,5 @@
     margin-bottom: 24px;
   }
 
-  .infoItem {
-    display: flex;
-    height: 50px;
-    border-bottom: 1px solid var(--border-color);
-    color: var(--paragraph-color);
-    align-items: center;
-  }
-
-  .infoItem label {
-    color: var(--text-gray);
-    margin-bottom: 8px;
-    width: 250px;
-    max-width: 30%;
-    flex: 0 0 auto;
-    display: inline-block;
-    margin: 0;
-  }
-
-  .infoValue {
-    font-size: 16px;
-  }
-
-  .afterCarouselLink {
-    font-size: 13px;
-    font-weight: 500;
-    color: var(--blue-500);
-    line-height: 1;
-    display: flex;
-    gap: 8px;
-    align-items: center;
-    transition: color 100ms ease-in-out;
-    margin-top: 24px;
-    margin-bottom: 40px;
-    
-    &:hover {
-      color: var(--blue-600);
-    }
-  }
-
-  .classesPlaceholder {
-    border: 1px solid var(--border-color);
-    border-radius: 8px;
-    background-color: var(--foreground-color);
-    height: 128px;
-  }
-
-  .schedulesPlaceholder {
-    border: 1px solid var(--border-color);
-    border-radius: 8px;
-    background-color: var(--foreground-color);
-    height: 84px;
-  }
+  
 }
