+++ apps/frontend/src/app/Profile/Support/index.tsx
@@ -1,0 +1,6 @@
+export default function Support() {
+  return <div>
+    <h1>Support</h1>
+    Thank you for beta testing our website!<br/>To leave us feedback, either use the blue button on the bottom right or email us at octo.berkeleytime@asuc.org.
+  </div>;
+}
