+++ apps/frontend/src/components/NavigationBar/index.tsx
@@ -60,13 +60,13 @@ export default function NavigationBar({ invert }: NavigationBarProps) {
             </MenuItem>
           )}
         </NavLink>
-        <NavLink to="/about">
+        {/* <NavLink to="/about">
           {({ isActive }) => (
             <MenuItem className={styles.item} active={isActive}>
               About
             </MenuItem>
           )}
-        </NavLink>
+        </NavLink> */}
       </div>
       {user ? (
         <Button
